requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    ./scripts/upgrade.sh
#
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.34.19
botocore==1.34.19
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
docker==7.0.0
fastapi==0.109.0
frozenlist==1.4.1
gitdb==4.0.11
gitpython==3.1.41
h11==0.14.0
haxor==1.2.4
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
mako==1.3.0
markdown-it-py==3.0.0
markupsafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.7.2
packaging==23.2
pandas==2.1.4
pgvector==0.2.4
phidata==2.3.6
pillow==10.2.0
pluggy==1.3.0
protobuf==4.25.2
psycopg[binary]==3.1.17
psycopg-binary==3.1.17
pyarrow==14.0.2
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-settings==2.1.0
pydeck==0.8.1b0
pygments==2.17.2
pypdf==3.17.4
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
ruff==0.1.13
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.25
starlette==0.35.1
streamlit==1.30.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typer==0.9.0
types-beautifulsoup4==4.12.0.20240106
types-html5lib==1.1.11.20240106
types-pillow==10.2.0.20240111
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.25.0
validators==0.22.0
yarl==1.9.4
zipp==3.17.0
